[scala] Implement Client

+++ scala-client/src/main/scala/com/drsquidop/scala_jruby_activerecord/scala_client/MyModel.scala
@@ -1,3 +1,3 @@
 package com.drsquidop.scala_jruby_activerecord_scala_client
 
-case class MyModel(id: Option[Int], name: String)
+case class MyModel(id: Option[Long], name: String)
